Commit migrated from https://github.com/mono/mono/commit/3e6c4f526499a339783cbe427e9916bbd99d7a0e

+++ src/mono/mono/mini/basic-float.cs
@@ -631,8 +631,8 @@ class Tests
 	public static int test_0_float_precision () {
 		float f1 = 3.40282346638528859E+38f;
 		float f2 = 3.40282346638528859E+38f;		
-		float PositiveInfinity =  1.0f / 0.0f;
-		float f = f1 + f2;
+		float PositiveInfinity =  (float)(1.0f / 0.0f);
+		float f = (float)(f1 + f2);
 
 		return f == PositiveInfinity ? 0 : 1;
 	}
